campaign_project/requirements.txt
Django==1.11.29
argparse==1.2.1
djangorestframework==3.9.1
pytz==2017.2
wsgiref==0.1.2
